[XPU] Fold LayerNorm into the layer_norm CustomOp
Replace the `nn.Module` `LayerNorm` with the renamed `layer_norm` CustomOp, deriving the compute dtype from `torch.promote_types(x.dtype, weight.dtype)` so fp32 parameters keep the old fp32 reduction.

+++ vllm/models/deepseek_v32/attention.py
@@ -84,7 +84,7 @@ class DeepseekV32Indexer(nn.Module):
             disable_tp=True,
             prefix=f"{prefix}.wk_weights_proj",
         )
-        self.k_norm = LayerNorm(self.head_dim, eps=1e-6)
+        self.k_norm = LayerNorm(self.head_dim, eps=1e-6, dtype=torch.float32)
         self.softmax_scale = self.head_dim**-0.5
 
         self.scale_fmt = "ue8m0"
